requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
PyPDF2==3.0.1
openai==1.10.0
python-dotenv==1.0.0
httpx==0.26.0
